Move delivery channel policy tests into a separate file
