+++ src/protagonist/API.Tests/Integration/CustomerPolicyTests.cs
@@ -93,37 +93,4 @@ public class CustomerPolicyTests : IClassFixture<ProtagonistAppFactory<Startup>>
         // Assert
         response.StatusCode.Should().Be(HttpStatusCode.NotFound);
     }
-    
-    [Fact]
-    public async Task Get_DeliveryChannelPolicy_200()
-    {
-        // Arrange
-        var path = $"customers/99/deliveryChannelPolicies/thumbs/example-thumbs-policy";
-  
-        // Act
-        var response = await httpClient.AsCustomer(99).GetAsync(path);
-        
-        // Assert
-        response.StatusCode.Should().Be(HttpStatusCode.OK);
-        
-        var model = await response.ReadAsHydraResponseAsync<DeliveryChannelPolicy>();
-        model.Name.Should().Be("example-thumbs-policy");
-        model.DisplayName.Should().Be("Example Thumbnail Policy");
-        model.Channel.Should().Be("thumbs");
-        model.PolicyData.Should().Be("{[\"!1024,1024\",\"!400,400\",\"!200,200\",\"!100,100\"]}");
-    }
-    
-    [Fact]
-    public async Task Get_DeliveryChannelPolicy_404_IfNotFound()
-    {
-        // Arrange
-        var path = $"customers/99/deliveryChannelPolicies/thumbs/foofoo";
-
-        // Act
-        var response = await httpClient.AsCustomer(99).GetAsync(path);
-
-        // Assert
-        response.StatusCode.Should().Be(HttpStatusCode.NotFound);
-    }
-    
 }
